Add files via upload
The goal of the Project is to explore reinforcement learning environments and implement actor-critic algorithms. In the rst part of the project will implement REINFORCE or (Monte Carlo PG) on Gym Acrobot Environment, and in the second part we will implement actor-critic(a2c) algorithm on Cartpole Environment. The purpose of this project is to understand the basic policy gradient algorithms. We will train our networks on a reinforcement learning environment among OpenAI Gym.